feat: Add generateOTP.js configuration file
Add a new file, generateOTP.js, to the backend/src/config directory. This file contains a function that generates a random OTP (One-Time Password) and returns it. The OTP is a 4-digit number generated using Math.random() and Math.floor(). If an error occurs during the generation process, it is thrown.

This commit adds the generateOTP.js file to the project, which will be used for generating OTPs in the future.